[refactor]apis로 객체화 로그아웃 구현

+++ src/components/Layout/Header/Header.js
@@ -3,18 +3,24 @@ import { flexCenter } from "libs/styles/common";
 import { TokenRepository } from "repository/TokenRepository";
 import styled from "styled-components";
 import {useNavigate} from 'react-router-dom'
+import { useState } from "react";
+
 
 function LayoutHeader(tokden) {
     const naviate = useNavigate();
-    
 
-    const onclickLogout = () =>{
-        const Removetoken = TokenRepository.getToken()
-        TokenRepository.removeToken(Removetoken);
-
-        if(!Removetoken){
-            naviate('/', { replace: true });
-            alert('로그아웃 되었습니다.')
+    const onclickLogout = async () =>{
+        try{
+        const res = await AuthApi.logout();
+        const Removetoken =  TokenRepository.getToken()
+        console.log(res);
+        if(res.status === 201){
+            TokenRepository.removeToken(Removetoken)
+                naviate('/', { replace: true });
+                alert('로그아웃 되었습니다.')
+            }
+        }catch(err) {
+            console.log(err);
         }
     }
 
@@ -22,7 +28,7 @@ function LayoutHeader(tokden) {
         <>
             <Wrapper>
                 <div>HEADER</div>
-                <button onClick={onclickLogout}>로그아웃</button>
+                    <button onClick={onclickLogout}>로그아웃</button>
             </Wrapper>
         </>
     )
